requirements.txt
Flask==3.0.2
Werkzeug==3.0.1 
anthropic==0.49.0
python-dotenv==1.0.1
openai==1.66.3